Fixed bug.
I was attempting to clean up some of the memory before by deleting a
value after it is printed. However, if the value was taken directly
from the AST this would corrupt the tree. For now, we will just not
clean up any memory. Down the road, reference counting will be needed
to clean up the memory.